@@ -1,6 +1,6 @@
 package com.viscompass.visompass;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.util.Date;
 
